prefer explicit MaterialX value types over addInputsFromNodeDef
This change is motivated by two reasons: first, the method adds ALL
inputs, instead of only the ones we want to set - polluting the
document. They are set to the default values anyway. Second, the method
adds the 'value' attribute, which we would have to remove in case of a
connection to a different node. In fact, we didn't do that, resulting in
overloaded inputs (with both 'value' and 'nodename' attributes). This is
illegal behavior and has recently been added to the MaterialX document
validation routine (although the recent MaterialX version is not
available in conjunction with USD, yet).