fix mtlx-as-usdshade errors with USD v22.11
Basically, we're importing the MaterialX libraries for verification of
the resulting document. However, when --mtlx-as-usdshade is requested,
then we hand over the document to UsdMtlx (more precisely, the UsdMtlxRead
function, which converts the document to a UsdShade network). It turns out
that in the recent USD version, this function respects 'imported'
NodeDefs, NodeGraphs and Materials, and tries to convert and export them,
but fails to do so. As a workaround, we only do validation when writing
an actual MaterialX document to a file.