Ifguard potentially unsupported MaterialX glTF PBR features
This is not strictly necessary as we add inputs regardless of the
provided NodeDef. However, an outdated NodeDef which is f.i. provided by
an older MaterialX version still produces document validation errors.

In order to avoid these errors, we make the hard assumption that the
provided MaterialX glTF PBR implementation is that of the corresponding
MaterialX C++ library, allowing us to conditionally disable newer and
unsupported features like KHR_materials_emissive_strength and
KHR_materials_iridescence.