Remove glTF PBR implementation option
Flattening was useful for compatibility with earlier MaterialX versions that
didn't ship with the glTF PBR. Today, we should expect the runtime to provide
the node, and if not, it is trivial to add. Another reason to get rid of
flattening is performance -- the number of nodes is much higher, and this
reduces usability because parsing is slow.

The portable glTF PBR implementation was an idea that did not work out due to
technical limitations. Unfortunately, the maintenance costs exceed the benefit
here, so we drop it.